trying to add crop image activity
