+++ app/src/main/java/com/good/scanner/kotlin/StillImageActivity.kt
@@ -18,6 +18,7 @@ package com.good.scanner.kotlin
 
 import android.app.Activity
 import android.content.ContentValues
+import android.content.Context
 import android.content.Intent
 import android.content.res.Configuration
 import android.graphics.Bitmap
@@ -55,7 +56,6 @@ import kotlin.math.sqrt
 /** Activity demonstrating different image detector features with a still image from camera.  */
 @KeepName
 class StillImageActivity : AppCompatActivity() {
-    private var preview: QuadrilateralSelectionImageView? = null
     private var graphicOverlay: com.good.scanner.GraphicOverlay? = null
     private var selectedMode = TEXT_RECOGNITION_KOREAN
     private var selectedSize: String? =
@@ -71,7 +71,6 @@ class StillImageActivity : AppCompatActivity() {
     private var imageProcessor: com.good.scanner.VisionImageProcessor? = null
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-
         setContentView(R.layout.activity_still_image)
         findViewById<View>(R.id.select_image_button)
                 .setOnClickListener { view: View ->
@@ -94,37 +93,6 @@ class StillImageActivity : AppCompatActivity() {
                     inflater.inflate(R.menu.camera_button_menu, popup.menu)
                     popup.show()
                 }
-        findViewById<View>(R.id.crop_image_button)
-                .setOnClickListener{ view: View? ->
-                    if (imageUri == null) return@setOnClickListener
-                    if (SIZE_SCREEN == selectedSize && imageMaxWidth == 0) return@setOnClickListener // UI layout has not finished yet, will reload once it's ready.
-                    val imageBitmap = getBitmapFromContentUri(contentResolver, imageUri) ?: return@setOnClickListener
-
-                    val points: List<PointF> = preview!!.getPoints()
-                    graphicOverlay!!.clear() // Clear the overlay first
-                    val resizedBitmap = getResizedBitmap(imageBitmap)
-                    val utils = ImgProcUtils()
-                    val orig = utils.convertBitmapToMat(resizedBitmap)
-                    // TODO: Image preprocessing - Perspective Transform
-                    val transformed = perspectiveTransform(orig, points)
-                    // TODO: Image preprocessing - Thresholding (Result image should be black & white)
-                    val mResult = applyThreshold(transformed!!)
-                    val resizedResultBitmap = getResizedBitmap(mResult!!)
-                    preview!!.deletePoints()
-                    preview!!.setImageBitmap(resizedResultBitmap)
-                    orig.release()
-                    transformed.release()
-
-                    // TODO: Run AI Model - Text Recognition
-                    if (imageProcessor != null) {
-                        graphicOverlay!!.setImageSourceInfo(resizedResultBitmap.width, resizedBitmap.height, false)
-                        imageProcessor!!.processBitmap(resizedResultBitmap, graphicOverlay)
-                    } else {
-                        Log.e(TAG, "Null imageProcessor, please check adb logs for imageProcessor creation error")
-                    }
-                }
-        preview = findViewById<QuadrilateralSelectionImageView>(R.id.preview)
-        graphicOverlay = findViewById(R.id.graphic_overlay)
 
         populateFeatureSelector()
         populateSizeSelector()
@@ -150,7 +118,7 @@ class StillImageActivity : AppCompatActivity() {
                         imageMaxHeight =
                                 rootView.height - findViewById<View>(R.id.control).height
                         if (com.good.scanner.kotlin.StillImageActivity.Companion.SIZE_SCREEN == selectedSize) {
-                            tryReloadAndDetectInImage()
+                            // tryReloadAndDetectInImage()
                         }
                     }
                 })
@@ -173,8 +141,8 @@ class StillImageActivity : AppCompatActivity() {
     public override fun onResume() {
         super.onResume()
         Log.d(com.good.scanner.kotlin.StillImageActivity.Companion.TAG, "onResume")
-        createImageProcessor()
-        tryReloadAndDetectInImage()
+        // createImageProcessor()
+        // tryReloadAndDetectInImage()
     }
 
     public override fun onPause() {
@@ -212,8 +180,8 @@ class StillImageActivity : AppCompatActivity() {
             ) {
                 if (pos >= 0) {
                     selectedMode = parentView.getItemAtPosition(pos).toString()
-                    createImageProcessor()
-                    tryReloadAndDetectInImage()
+                    // createImageProcessor()
+                    // tryReloadAndDetectInImage()
                 }
             }
 
@@ -244,7 +212,7 @@ class StillImageActivity : AppCompatActivity() {
             ) {
                 if (pos >= 0) {
                     selectedSize = parentView.getItemAtPosition(pos).toString()
-                    tryReloadAndDetectInImage()
+                    // tryReloadAndDetectInImage()
                 }
             }
 
@@ -262,7 +230,6 @@ class StillImageActivity : AppCompatActivity() {
 
     private fun startCameraIntentForResult() { // Clean up last time's image
         imageUri = null
-        preview!!.setImageBitmap(null)
         val takePictureIntent = Intent(MediaStore.ACTION_IMAGE_CAPTURE)
         if (takePictureIntent.resolveActivity(packageManager) != null) {
             val values = ContentValues()
@@ -287,91 +254,15 @@ class StillImageActivity : AppCompatActivity() {
             data: Intent?
     ) {
         if (requestCode == REQUEST_IMAGE_CAPTURE && resultCode == Activity.RESULT_OK) {
-            tryReloadAndDetectInImage()
+            // TODO: Intent를 이용하여 다른 Activity로 image uri 넘기기
+            // startCropImageActivity()
         } else if (requestCode == REQUEST_CHOOSE_IMAGE && resultCode == Activity.RESULT_OK) {
             // In this case, imageUri is returned by the chooser, save it.
             imageUri = data!!.data
-            tryReloadAndDetectInImage()
+            // TODO: Intent를 이용하여 다른 Activity로 image uri 넘기기
+            // startCropImageActivity()
         } else {
             super.onActivityResult(requestCode, resultCode, data)
-        }
-    }
-
-    private fun getResizedBitmap(bitmap: Bitmap): Bitmap {
-        val scaleFactor = (bitmap.width.toFloat() / targetedWidthHeight.first.toFloat())
-                .coerceAtLeast(bitmap.height.toFloat() / targetedWidthHeight.second.toFloat())
-        return when(selectedSize == SIZE_ORIGINAL) {
-            true -> bitmap
-            false -> Bitmap.createScaledBitmap(
-                bitmap,
-                (bitmap.width / scaleFactor).toInt(),
-                (bitmap.height / scaleFactor).toInt(),
-                true
-            )
-        }
-    }
-
-    private fun tryReloadAndDetectInImage() {
-        Log.d(TAG, "Try reload and detect image")
-        try {
-            if (imageUri == null) return
-            if (SIZE_SCREEN == selectedSize && imageMaxWidth == 0) return // UI layout has not finished yet, will reload once it's ready.
-
-            val imageBitmap = getBitmapFromContentUri(contentResolver, imageUri) ?: return
-            graphicOverlay!!.clear() // Clear the overlay first
-            val resizedBitmap = getResizedBitmap(imageBitmap)
-
-            // TODO:
-            //  image = imutils.resize(image, height = 500)
-            //  coef_y = orig.shape[0] / image.shape[0]
-            //  coef_x = orig.shape[1] / image.shape[1]
-            //  screenCnt[:, :, 0] = screenCnt[:, :, 0] * coef_x
-            //  screenCnt[:, :, 1] = screenCnt[:, :,  1] * coef_y
-
-            val scaleFactor = (resizedBitmap.height.toFloat() / 500.0) // Height가 500이 되도록 image를 rescaling. edge detection에는 낮은 해상도의 이미지가 더 제격이다
-            val scaledBitmap = Bitmap.createScaledBitmap(
-                resizedBitmap,
-                (resizedBitmap.width / scaleFactor).toInt(),
-                (resizedBitmap.height / scaleFactor).toInt(),
-                true
-            )
-            preview!!.setImageBitmap(resizedBitmap)
-
-            val scaleY = resizedBitmap.height.toFloat() / scaledBitmap.height.toFloat()
-            val scaleX = resizedBitmap.width.toFloat() / scaledBitmap.width.toFloat()
-
-            val utils = ImgProcUtils()
-            val imageMat = utils.convertBitmapToMat(resizedBitmap)
-            val scaledImageMat = utils.convertBitmapToMat(scaledBitmap)
-            val greyMat = utils.convertToGreyScale(scaledImageMat)
-            val blurredMat = utils.convertToBlurred(greyMat)
-            val edgesMat = utils.convertToEdgeDetected(blurredMat)
-            val contours = utils.findContours(edgesMat)
-
-            contours.firstOrNull {
-                utils.approximatePolygonal(it).toArray().size == 4
-            }?.let {
-                Log.d(TAG,"Contour Found!")
-                val scaledContour = utils.scaleContour(imageMat, it, scaleX, scaleY)
-                val result = ArrayList<PointF>()
-                val points = sortPoints(scaledContour.toArray())
-                result.add(PointF(java.lang.Double.valueOf(points[0].x).toFloat(), java.lang.Double.valueOf(points[0].y).toFloat()))
-                result.add(PointF(java.lang.Double.valueOf(points[1].x).toFloat(), java.lang.Double.valueOf(points[1].y).toFloat()))
-                result.add(PointF(java.lang.Double.valueOf(points[2].x).toFloat(), java.lang.Double.valueOf(points[2].y).toFloat()))
-                result.add(PointF(java.lang.Double.valueOf(points[3].x).toFloat(), java.lang.Double.valueOf(points[3].y).toFloat()))
-                preview?.setPoints(result)
-            } ?: {
-                preview?.setPoints(null)
-            }
-            edgesMat.release()
-            blurredMat.release()
-            greyMat.release()
-            scaledImageMat.release()
-            imageMat.release()
-
-        } catch (e: IOException) {
-            Log.e(TAG, "Error retrieving saved image")
-            imageUri = null
         }
     }
 
@@ -397,123 +288,6 @@ class StillImageActivity : AppCompatActivity() {
             return Pair(targetWidth, targetHeight)
         }
 
-    private fun createImageProcessor() {
-        try {
-            when (selectedMode) {
-                TEXT_RECOGNITION_KOREAN ->
-                    imageProcessor =
-                            TextRecognitionProcessor(this, KoreanTextRecognizerOptions.Builder().build())
-                else -> Log.e(TAG, "Unknown selectedMode: $selectedMode")
-            }
-        } catch (e: Exception) {
-            Log.e(TAG, "Can not create image processor: $selectedMode", e)
-            Toast.makeText(
-                    applicationContext,
-                    "Can not create image processor: " + e.message,
-                    Toast.LENGTH_LONG
-            )
-                    .show()
-        }
-    }
-
-    /**
-     * Transform the coordinates on the given Mat to correct the perspective.
-     *
-     * @param src A valid Mat
-     * @param points A list of coordinates from the given Mat to adjust the perspective
-     * @return A perspective transformed Mat
-     */
-    private fun perspectiveTransform(src: Mat, points: List<PointF>): Mat? {
-        val point1 = Point(points[0].x.toDouble(), points[0].y.toDouble())
-        val point2 = Point(points[1].x.toDouble(), points[1].y.toDouble())
-        val point3 = Point(points[2].x.toDouble(), points[2].y.toDouble())
-        val point4 = Point(points[3].x.toDouble(), points[3].y.toDouble())
-        val pts = arrayOf(point1, point2, point3, point4)
-        return fourPointTransform(src, sortPoints(pts))
-    }
-
-    /**
-     * Apply a threshold to give the "scanned" look
-     *
-     * NOTE:
-     * See the following link for more info http://docs.opencv.org/3.1.0/d7/d4d/tutorial_py_thresholding.html#gsc.tab=0
-     * @param src A valid Mat
-     * @return The processed Bitmap
-     */
-    private fun applyThreshold(src: Mat): Bitmap? {
-        Imgproc.cvtColor(src, src, Imgproc.COLOR_BGR2GRAY)
-
-        // Some other approaches
-//        Imgproc.adaptiveThreshold(src, src, 255, Imgproc.ADAPTIVE_THRESH_MEAN_C, Imgproc.THRESH_BINARY, 15, 15);
-//        Imgproc.threshold(src, src, 0, 255, Imgproc.THRESH_BINARY + Imgproc.THRESH_OTSU);
-        Imgproc.GaussianBlur(src, src, Size(5.0, 5.0), 0.0)
-        Imgproc.adaptiveThreshold(src, src, 255.0, Imgproc.ADAPTIVE_THRESH_GAUSSIAN_C, Imgproc.THRESH_BINARY, 11, 2.0)
-        val bm = Bitmap.createBitmap(src.width(), src.height(), Bitmap.Config.ARGB_8888)
-        Utils.matToBitmap(src, bm)
-        return bm
-    }
-
-    /**
-     * Sort the points
-     *
-     * The order of the points after sorting:
-     * 0------->1
-     * ^        |
-     * |        v
-     * 3<-------2
-     *
-     * NOTE:
-     * Based off of http://www.pyimagesearch.com/2014/08/25/4-point-opencv-getperspective-transform-example/
-     *
-     * @param src The points to sort
-     * @return An array of sorted points
-     */
-    private fun sortPoints(src: Array<Point>): Array<Point> {
-        val srcPoints = ArrayList(listOf(*src))
-        val result = arrayOf<Point>(Point(0.0,0.0), Point(0.0,0.0), Point(0.0,0.0), Point(0.0,0.0))
-
-        val sumComparator  = compareBy<Point> {it.y + it.x}
-        val differenceComparator =compareBy<Point> {it.y - it.x}
-        result[0] = Collections.min(srcPoints, sumComparator) // Upper left has the minimal sum
-        result[2] = Collections.max(srcPoints, sumComparator) // Lower right has the maximal sum
-        result[1] = Collections.min(srcPoints, differenceComparator) // Upper right has the minimal difference
-        result[3] = Collections.max(srcPoints, differenceComparator) // Lower left has the maximal difference
-        return result
-    }
-
-    /**
-     * NOTE:
-     * Based off of http://www.pyimagesearch.com/2014/08/25/4-point-opencv-getperspective-transform-example/
-     *
-     * @param src
-     * @param pts
-     * @return
-     */
-    private fun fourPointTransform(src: Mat, pts: Array<Point>): Mat {
-        val ratio = 1.0
-        val ul = pts[0]
-        val ur = pts[1]
-        val lr = pts[2]
-        val ll = pts[3]
-        val widthA = sqrt((lr.x - ll.x).pow(2.0) + (lr.y - ll.y).pow(2.0))
-        val widthB = sqrt((ur.x - ul.x).pow(2.0) + (ur.y - ul.y).pow(2.0))
-        val maxWidth = Math.max(widthA, widthB) * ratio
-        val heightA = sqrt((ur.x - lr.x).pow(2.0) + (ur.y - lr.y).pow(2.0))
-        val heightB = sqrt((ul.x - ll.x).pow(2.0) + (ul.y - ll.y).pow(2.0))
-        val maxHeight = Math.max(heightA, heightB) * ratio
-        val resultMat = Mat(java.lang.Double.valueOf(maxHeight).toInt(), java.lang.Double.valueOf(maxWidth).toInt(), CvType.CV_8UC4)
-        val srcMat = Mat(4, 1, CvType.CV_32FC2)
-        val dstMat = Mat(4, 1, CvType.CV_32FC2)
-        srcMat.put(0, 0, ul.x * ratio, ul.y * ratio, ur.x * ratio, ur.y * ratio, lr.x * ratio, lr.y * ratio, ll.x * ratio, ll.y * ratio)
-        dstMat.put(0, 0, 0.0, 0.0, maxWidth, 0.0, maxWidth, maxHeight, 0.0, maxHeight)
-        val M = Imgproc.getPerspectiveTransform(srcMat, dstMat)
-        Imgproc.warpPerspective(src, resultMat, M, resultMat.size())
-        srcMat.release()
-        dstMat.release()
-        M.release()
-        return resultMat
-    }
-
     companion object {
         private const val TAG = "StillImageActivity"
         private const val TEXT_RECOGNITION_KOREAN = "Text Recognition Korean"
@@ -528,6 +302,7 @@ class StillImageActivity : AppCompatActivity() {
         private const val KEY_SELECTED_SIZE = "com.google.mlkit.vision.demo.KEY_SELECTED_SIZE"
         private const val REQUEST_IMAGE_CAPTURE = 1001
         private const val REQUEST_CHOOSE_IMAGE = 1002
+        const val IMAGE_URI = "IMAGE_URI"
 
         init {
             if (!OpenCVLoader.initDebug()) {
