+++ app/src/main/res/layout/activity_still_image.xml
@@ -51,14 +51,6 @@
           android:layout_margin="12dp"
           android:text="@string/select_image"/>
 
-      <Button
-          android:id="@+id/crop_image_button"
-          android:layout_width="wrap_content"
-          android:layout_height="wrap_content"
-          android:layout_gravity="center"
-          android:layout_margin="12dp"
-          android:text="@string/crop_image" />
-
     </LinearLayout>
     <LinearLayout
         android:id="@+id/control2"
